+++ tests/models/test_anthropic_thinking_blocks.py
