+++ tests/models/test_extended_thinking_message_order.py
