+++ tests/models/test_gemini_thought_signatures_stream.py
