+++ tests/models/test_openai_chatcompletions.py
