+++ tests/models/test_openai_chatcompletions_converter.py
