+++ tests/models/test_openai_chatcompletions_stream.py
