+++ tests/models/test_openai_client_utils.py
