+++ tests/models/test_openai_responses.py
