+++ tests/models/test_openai_responses_converter.py
