+++ tests/models/test_reasoning_content.py
