+++ tests/models/test_remove_openai_responses_api_incompatible_fields.py
