+++ tests/realtime/test_session_exceptions.py
